+++ test/cases/entry-data-i18n-multipage/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const HtmlBundlerPlugin = require('../../../');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
 
 const languages = {
   'de-DE': require('./src/locales/de-DE.json'),