+++ test/cases/integrity-publicPath-root-complex/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const HtmlBundlerPlugin = require('../../../');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
 
 module.exports = {
   //mode: 'development',