+++ test/cases/js-import-css-same-in-many-mini-css/expected/news.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";var o={743:()=>{console.log(">> main-a"),console.log(">> app"),console.log(">> home")}},e={};function r(s){var n=e[s];if(void 0!==n)return n.exports;var t=e[s]={exports:{}};return o[s](t,t.exports,r),t.exports}r(743),console.log(">> news")})();