+++ test/cases/js-import-css-same-in-many-mini-css/src/app.js
@@ -1,0 +1,3 @@
+import './main-a';
+
+console.log('>> app');