+++ test/cases/js-import-css-same-in-many-mini-css/src/home.js
@@ -1,0 +1,6 @@
+import './home.css';
+import './common.css';
+
+import './app';
+
+console.log('>> home');