+++ test/cases/js-import-css-same-in-many-mini-css/src/news.js
@@ -1,0 +1,6 @@
+import './common.css';
+import './news.css';
+
+import './home';
+
+console.log('>> news');