+++ test/cases/js-import-css-same-in-many-mini-css/webpack.config.js
@@ -1,0 +1,26 @@
+const path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  entry: {
+    home: './src/home.js',
+    news: './src/news.js',
+  },
+
+  plugins: [new MiniCssExtractPlugin()],
+
+  module: {
+    rules: [
+      {
+        test: /\.(css)/,
+        use: [MiniCssExtractPlugin.loader, 'css-loader'],
+      },
+    ],
+  },
+};