+++ test/cases/msg-exception-handlebars-include/webpack.config.js
@@ -1,4 +1,4 @@
-const HtmlBundlerPlugin = require('../../../');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
 
 module.exports = {
   mode: 'production',