+++ test/cases/options-minimal/webpack.config.js
@@ -1,4 +1,4 @@
-const HtmlBundlerPlugin = require('../../../');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
 
 // test minimal Webpack options
 module.exports = {