+++ test/cases/resolve-in-many-pages-from-same-tmpl/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const HtmlBundlerPlugin = require('../../../');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
 
 const render = (content, data) => {
   for (const key in data) {